General updates (#5)
This PR aims to make the template easier to use for those new to Jekyll and the Just the Docs theme. The changes in this PR are merely my suggestions (based on my own lack of understanding of what terms like "deploy" really mean).

- Updates the home page at `index.md` to give an overview of *what* the template can do.
- Updates `README.md` to focus on *how* to do it ([preview](https://github.com/pdmosses/just-the-docs-template/blob/general-updates/README.md)).
- Minor polishing of `.gitignore`, `_config.yml`, and `Gemfile`.

When the template has been finalised, I'd like to add a reference to it on the theme home page.