Fix missing docs step (enable GitHub Actions as deploy source) (#32)
I've been meaning to do this for a while, but haven't gotten around to it! This PR adds the missing step of having to explicitly enable GitHub Actions as the build source for GitHub Pages.

Open to suggestions on better copy!

This closes #17 (and various other issues/discussions about this template not working properly).